BUG/MAJOR: http: fix again http-send-name-header
Several months ago, Guillaume Castagnino reported another issue in
relation with http-send-name-header in 1.4 when the requests are
pipelined. Several attempts at fixing this failed and some of them
could even lead the process to crash upon posts, as Cyril's showed.

It turns out we were looking in the wrong direction as the problem
is directly related to the header addition and nothing is related to
the forwarding.

The cause of the problem is the way pointers to the beginning of the
body are found during the forwarding. msg->som is used there because
it points to the beginning of the body. This pointer was not updated
during the header addition because it was unexpected to have to do
this since we're already in the data phase.

This patch fixes both msg->som and msg->eol which are the only two
values which used to point to the end of headers before the addition
and which might be expected to still point there after addition. It
is not certain ->eol is needed at this stage but since the code is
tricky, better safe than sorry.

Now pipelined requests correctly work, redispatch with names of
different sizes still works, and balancing on POST also works, even
after a redispatch.

This patch is specific to version 1.4, the feature does not exist
in older versions and the buffer/messages are arranged differently
in later ones.