BUG/MAJOR: buffer: fix possible integer overflow on reserved size computation
This is patch does the equivalent of the following commit introduced
into 1.6. In 1.4 the code is completely different but this fix was
thoroughly tested.

---------
From 0428a146c0ee776f7c89359e8aed0d755ffbad07 Mon Sep 17 00:00:00 2001
From: Willy Tarreau <w@1wt.eu>
Date: Thu, 8 Jan 2015 11:34:55 +0100
Subject: [PATCH] BUG/MEDIUM: channel: fix possible integer overflow on
 reserved size computation

The buffer_max_len() function is subject to an integer overflow in this
calculus :

    int ret = global.tune.maxrewrite - chn->to_forward - chn->buf->o;

  - chn->to_forward may be up to 2^31 - 1
  - chn->buf->o may be up to chn->buf->size
  - global.tune.maxrewrite is by definition smaller than chn->buf->size

Thus here we can subtract (2^31 + buf->o) (highly negative) from something
slightly positive, and result in ret being larger than expected.

Fortunately in 1.5 and 1.6, this is only used by bi_avail() which itself
is used by applets which do not set high values for to_forward so this
problem does not happen there. However in 1.4 the equivalent computation
was used to limit the size of a read and can result in a read overflow
when combined with the nasty http-send-name-header feature.