BUG/MAJOR: buffer: don't schedule data in transit for leaving until connected
This is patch does the equivalent of the following series of commits
introduced into 1.6 :

bb3f994 BUG/MINOR: channel: compare to_forward with buf->i, not buf->size
1a4484d MINOR: channel: add channel_in_transit()
fe57834 MEDIUM: channel: make buffer_reserved() use channel_in_transit()
27bb0e1 MEDIUM: channel: make bi_avail() use channel_in_transit()
9c06ee4 BUG/MEDIUM: channel: don't schedule data in transit for leaving until connected

The most important one is this last one whose detailed commit is reproduced
below. In 1.4, the code is quite different and a lot simpler, so in practice
it's sufficient to add the control to buffer_max_len(). There's no need to
backport this fix to 1.3 as it only affects the http-send-name-header feature
which was introduced late in 1.4.

-----------
From 9c06ee4ccfded412739fb94260ba105be65651f8 Mon Sep 17 00:00:00 2001
From: Willy Tarreau <w@1wt.eu>
Date: Wed, 14 Jan 2015 16:08:45 +0100
Subject: [PATCH] BUG/MEDIUM: channel: don't schedule data in transit for
 leaving until connected

Option http-send-name-header is still hurting. If a POST request has to be
redispatched when this option is used, and the next server's name is larger
than the initial one, and the POST body fills the buffer, it becomes
impossible to rewrite the server's name in the buffer when redispatching.
In 1.4, this is worse, the process may crash because of a negative size
computation for the memmove().

The only solution to fix this is to refrain from eating the reserve before
we're certain that we won't modify the buffer anymore. And the condition for
that is that the connection is established.

This patch introduces "channel_may_send()" which helps to detect whether it's
safe to eat the reserve or not. This condition is used by channel_in_transit()
introduced by recent patches.

This patch series must be backported into 1.5, and a simpler version must be
backported into 1.4 where fixing the bug is much easier since there were no
channels by then. Note that in 1.4 the severity is major.