BUG/MEDIUM: http: incorrect transfer-coding in the request is a bad request
The rules related to how to handle a bad transfer-encoding header (one
where "chunked" is not at the final place) have evolved to mandate an
abort when this happens in the request. Previously it was only a close
(which is still valid for the server side).

This must be backported to 1.5 and 1.4.
(cherry picked from commit 34dfc605717e6e196a2a8553254437fcc4844cbc)
(cherry picked from commit 49efa267a15fdea781e5e78d2496c1130d186c54)